requirements.txt
reflex
supabase
requests
python-dotenv

supabase
python-dotenv
requests
fastapi
uvicorn
pydantic
reflex==0.8.6